+++ src/dawn/native/CommandEncoder.cpp
@@ -1230,7 +1230,7 @@ namespace dawn::native {
             [&](CommandAllocator* allocator) -> MaybeError {
                 InsertDebugMarkerCmd* cmd =
                     allocator->Allocate<InsertDebugMarkerCmd>(Command::InsertDebugMarker);
-                cmd->length = strlen(groupLabel);
+                cmd->length = static_cast<uint32_t>(strlen(groupLabel));
 
                 char* label = allocator->AllocateData<char>(cmd->length + 1);
                 memcpy(label, groupLabel, cmd->length + 1);
@@ -1264,7 +1264,7 @@ namespace dawn::native {
             [&](CommandAllocator* allocator) -> MaybeError {
                 PushDebugGroupCmd* cmd =
                     allocator->Allocate<PushDebugGroupCmd>(Command::PushDebugGroup);
-                cmd->length = strlen(groupLabel);
+                cmd->length = static_cast<uint32_t>(strlen(groupLabel));
 
                 char* label = allocator->AllocateData<char>(cmd->length + 1);
                 memcpy(label, groupLabel, cmd->length + 1);
