+++ src/dawn/native/vulkan/BufferVk.cpp
@@ -224,7 +224,7 @@ namespace dawn::native::vulkan {
 
         // Initialize the padding bytes to zero.
         if (device->IsToggleEnabled(Toggle::LazyClearResourceOnFirstUse) && !mappedAtCreation) {
-            uint32_t paddingBytes = GetAllocatedSize() - GetSize();
+            uint32_t paddingBytes = static_cast<uint32_t>(GetAllocatedSize() - GetSize());
             if (paddingBytes > 0) {
                 uint32_t clearSize = Align(paddingBytes, 4);
                 uint64_t clearOffset = GetAllocatedSize() - clearSize;
